fastapi_auth/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
greenlet>=3.0.0

# Authentication
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6

# Rate limiting
slowapi==0.1.9

# Environment variables
python-dotenv==1.0.0

# Email validation
email-validator==2.1.0
